change to SDL3, seg fault in SDL_Quit()

+++ omega/events/input_manager.cpp
@@ -17,17 +17,17 @@ bool InputManager::poll_events(Event &event) {
 void InputManager::update() {
     key_manager.update();
     // get the mouse state and store the previous frame's state
-    int x, y;
+    f32 x, y;
     mouse.prev_buttons = mouse.buttons;
     mouse.last_pos = mouse.pos;
     if (mouse.relative_mode) {
         mouse.buttons = SDL_GetRelativeMouseState(&x, &y);
-        mouse.pos.x = (f32)x;
-        mouse.pos.y = -(f32)y;
+        mouse.pos.x = x;
+        mouse.pos.y = -y;
     } else {
         mouse.buttons = SDL_GetMouseState(&x, &y);
-        mouse.pos.x = (f32)x;
-        mouse.pos.y = (f32)(core::Window::instance()->get_height() - y);
+        mouse.pos.x = x;
+        mouse.pos.y = (core::Window::instance()->get_height() - y);
     }
 }
